Introduce BYO cluster. Add two new tests to test_send, testing simple save and send execution on byo cluster, and a call variants test. Send test passes, reloading passes (but must be refactored to put ssh_creds in Secrets), call variant test fails on retrieving remoted ref (pid_fn.get(pid_ref)).
