+++ runhouse/rns/hardware/skycluster.py
@@ -1,27 +1,19 @@
-from typing import List, Optional, Union
 import subprocess
 import logging
 from pathlib import Path
-import pkgutil
 import contextlib
 import yaml
 
 import sky
 from sky.backends import backend_utils, CloudVmRayBackend
-from sky.utils import command_runner
-from sshtunnel import SSHTunnelForwarder, HandlerSSHTunnelForwarderError
-import ray.cloudpickle as pickle
 
-from runhouse.rns.resource import Resource
-from runhouse.rns.packages.package import Package
-from runhouse.grpc_handler.unary_client import UnaryClient
-from runhouse.grpc_handler.unary_server import UnaryService
-from runhouse.rh_config import rns_client, configs, open_grpc_tunnels
+from runhouse.rns.hardware.cluster import Cluster
+from runhouse.rh_config import rns_client, configs
 
 logger = logging.getLogger(__name__)
 
 
-class Cluster(Resource):
+class SkyCluster(Cluster):
     RESOURCE_TYPE = "cluster"
     DEFAULT_AUTOSTOP_MINS = 10
 
@@ -34,6 +26,7 @@ class Cluster(Resource):
                  autostop_mins=None,
                  use_spot=False,
                  image_id=None,
+                 region=None,
                  sky_data=None,
                  **kwargs  # We have this here to ignore extra arguments when calling from from_config
                  ):
@@ -57,6 +50,7 @@ class Cluster(Resource):
             else configs.get('default_autostop')
         self.use_spot = use_spot if use_spot is not None else configs.get('use_spot')
         self.image_id = image_id
+        self.region = region
 
         self.address = None
         self._yaml_path = None
@@ -76,7 +70,7 @@ class Cluster(Resource):
 
     @staticmethod
     def from_config(config: dict, dryrun=False):
-        return Cluster(**config, dryrun=dryrun)
+        return SkyCluster(**config, dryrun=dryrun)
 
     @property
     def config_for_rns(self):
@@ -89,6 +83,7 @@ class Cluster(Resource):
                        'autostop_mins': self.autostop_mins,
                        'use_spot': self.use_spot,
                        'image_id': self.image_id,
+                       'region': self.region,
                        'sky_data': self._get_sky_data(),
                        })
         return config
@@ -231,12 +226,10 @@ class Cluster(Resource):
         return self
 
     def up(self,
-           region: str = None,
            ):
         if self.provider in ['aws', 'gcp', 'azure', 'cheapest']:
-            task = sky.Task(  # run=f'echo SkyPilot cluster {self.name} launched.',
+            task = sky.Task(
                 num_nodes=self.num_instances if ':' not in self.instance_type else None,
-                # workdir=package,
                 # docker_image=image,  # Zongheng: this is experimental, don't use it
                 # envs=None,
             )
@@ -247,7 +240,7 @@ class Cluster(Resource):
                     cloud=cloud_provider,
                     instance_type=self.instance_type if ':' not in self.instance_type else None,
                     accelerators=self.instance_type if ':' in self.instance_type else None,
-                    region=region,  # TODO
+                    region=self.region,
                     image_id=self.image_id,
                     use_spot=self.use_spot
                 )
@@ -274,71 +267,6 @@ class Cluster(Resource):
         self.populate_vars_from_status()
         self.restart_grpc_server()
 
-    def sync_runhouse_to_cluster(self, _install_url=None):
-        if not self.address:
-            raise ValueError(f'No address set for cluster <{self.name}>. Is it up?')
-        local_rh_package_path = Path(pkgutil.get_loader('runhouse').path).parent
-
-        # Check if runhouse is installed from source and has setup.py
-        if not _install_url and \
-                local_rh_package_path.parent.name == 'runhouse' and \
-                (local_rh_package_path.parent / 'setup.py').exists():
-            # Package is installed in editable mode
-            local_rh_package_path = local_rh_package_path.parent
-            rh_package = Package.from_string(f'reqs:{local_rh_package_path}', dryrun=True)
-            rh_package.to_cluster(self, mount=False)
-            status_codes = self.run(['pip install ./runhouse'], stream_logs=True)
-        # elif local_rh_package_path.parent.name == 'site-packages':
-        else:
-            # Package is installed in site-packages
-            # status_codes = self.run(['pip install runhouse-nightly==0.0.2.20221202'], stream_logs=True)
-            # rh_package = 'runhouse_nightly-0.0.1.dev20221202-py3-none-any.whl'
-            # rh_download_cmd = f'curl https://runhouse-package.s3.amazonaws.com/{rh_package} --output {rh_package}'
-            # TODO need to check user's current version and install same version?
-            _install_url = _install_url or 'runhouse'
-            rh_install_cmd = f'pip install {_install_url}'
-            status_codes = self.run([rh_install_cmd], stream_logs=True)
-
-        if status_codes[0][0] != 0:
-            raise ValueError(f'Error installing runhouse on cluster <{self.name}>')
-
-    def install_packages(self, reqs: List[Union[Package, str]]):
-        if not self.is_connected():
-            self.connect_grpc()
-        to_install = []
-        # TODO [DG] validate package strings
-        for package in reqs:
-            if isinstance(package, str):
-                # If the package is a local folder, we need to create the package to sync it over to the cluster
-                pkg_obj = Package.from_string(package, dryrun=False)
-            else:
-                pkg_obj = package
-
-            from runhouse.rns.folders.folder import Folder
-            if isinstance(pkg_obj.install_target, Folder) and \
-                    pkg_obj.install_target.is_local():
-                pkg_str = pkg_obj.name or Path(pkg_obj.install_target.url).name
-                logging.info(f'Copying local package {pkg_str} to cluster <{self.name}>')
-                remote_package = pkg_obj.to_cluster(self, mount=False, return_dest_folder=True)
-                to_install.append(remote_package)
-            else:
-                to_install.append(package)  # Just appending the string!
-        # TODO replace this with figuring out how to stream the logs when we install
-        logging.info(f'Installing packages on cluster {self.name}: '
-                     f'{[req if isinstance(req, str) else str(req) for req in reqs]}')
-        self.client.install_packages(pickle.dumps(to_install))
-
-    def get(self, key, default=None, stream_logs=False):
-        if not self.is_connected():
-            self.connect_grpc()
-        return self.client.get_object(key, stream_logs=stream_logs) or default
-
-    def clear_pins(self, pins: Optional[List[str]] = None):
-        if not self.is_connected():
-            self.connect_grpc()
-        self.client.clear_pins(pins)
-        logger.info(f'Clearing pins on cluster {pins or ""}')
-
     def keep_warm(self, autostop_mins=-1):
         sky.autostop(self.name, autostop_mins, down=True)
         self.autostop_mins = autostop_mins
@@ -352,176 +280,13 @@ class Cluster(Resource):
         self.teardown()
         rns_client.delete_configs()
 
-    # ----------------- gRPC Methods ----------------- #
-
-    def connect_grpc(self, force_reconnect=False):
-        # FYI based on: https://sshtunnel.readthedocs.io/en/latest/#example-1
-        # FYI If we ever need to do this from scratch, we can use this example:
-        # https://github.com/paramiko/paramiko/blob/main/demos/rforward.py#L74
-        if not self.address:
-            raise ValueError(f'No address set for cluster <{self.name}>. Is it up?')
-
-        # TODO [DG] figure out how to ping to see if tunnel is already up
-        if self._grpc_tunnel and force_reconnect:
-            self._grpc_tunnel.close()
-
-        # TODO Check if port is already open instead of refcounting?
-        # status = subprocess.run(['nc', '-z', self.address, str(self.grpc_port)], capture_output=True)
-        # if not self.check_port(self.address, UnaryClient.DEFAULT_PORT):
-
-        tunnel_refcount = 0
-        if self.address in open_grpc_tunnels:
-            ssh_tunnel, connected_port, tunnel_refcount = open_grpc_tunnels[self.address]
-            ssh_tunnel.check_tunnels()
-            if ssh_tunnel.tunnel_is_up[ssh_tunnel.local_bind_address]:
-                self._grpc_tunnel = ssh_tunnel
-        else:
-            self._grpc_tunnel, connected_port = self.ssh_tunnel(UnaryClient.DEFAULT_PORT,
-                                                                remote_port=UnaryService.DEFAULT_PORT,
-                                                                num_ports_to_try=5)
-        open_grpc_tunnels[self.address] = (self._grpc_tunnel, connected_port, tunnel_refcount + 1)
-
-        # Connecting to localhost because it's tunneled into the server at the specified port.
-        self.client = UnaryClient(host='127.0.0.1', port=connected_port)
-
-    @staticmethod
-    def check_port(ip_address, port):
-        import socket
-        s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        return s.connect_ex(('127.0.0.1', int(port)))
-
-    def ssh_tunnel(self,
-                   local_port,
-                   remote_port=None,
-                   num_ports_to_try: int = 0) -> (SSHTunnelForwarder, int):
-        # Debugging cmds (mac):
-        # netstat -vanp tcp | grep 5005
-        # lsof -i :5005_
-        # kill -9 <pid>
-
-        creds: dict = self.ssh_creds()
-        connected = False
-        ssh_tunnel = None
-        while not connected:
-            try:
-                if local_port > local_port + num_ports_to_try:
-                    raise Exception(f'Failed to create SSH tunnel after {num_ports_to_try} attempts')
-
-                ssh_tunnel = SSHTunnelForwarder(
-                    self.address,
-                    ssh_username=creds['ssh_user'],
-                    ssh_pkey=creds['ssh_private_key'],
-                    local_bind_address=('', local_port),
-                    remote_bind_address=('127.0.0.1', remote_port or local_port),
-                    set_keepalive=1,
-                    # mute_exceptions=True,
-                )
-                ssh_tunnel.start()
-                connected = True
-            except HandlerSSHTunnelForwarderError as e:
-                # try connecting with a different port - most likely the issue is the port is already taken
-                local_port += 1
-                pass
-
-        return ssh_tunnel, local_port
-
-    # TODO [DG] Remove this for now, for some reason it was causing execution to hang after programs completed
-    # def __del__(self):
-    # if self.address in open_grpc_tunnels:
-    #     tunnel, port, refcount = open_grpc_tunnels[self.address]
-    #     if refcount == 1:
-    #         tunnel.stop(force=True)
-    #         open_grpc_tunnels.pop(self.address)
-    #     else:
-    #         open_grpc_tunnels[self.address] = (tunnel, port, refcount - 1)
-    # elif self._grpc_tunnel:  # Not sure why this would be reached but keeping it just in case
-    #     self._grpc_tunnel.stop(force=True)
-
-    # import paramiko
-    # ssh = paramiko.SSHClient()
-    # ssh.load_system_host_keys()
-    # ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-    # from pathlib import Path
-    # ssh.connect(self.address,
-    #             username=creds['ssh_user'],
-    #             key_filename=str(Path(creds['ssh_private_key']).expanduser())
-    #             )
-    # transport = ssh.get_transport()
-    # transport.request_port_forward('', local_port)
-    # ssh_tunnel = transport.open_channel("direct-tcpip", ("localhost", local_port),
-    #                                     (self.address, remote_port or local_port))
-    # if ssh_tunnel.is_active():
-    #     connected = True
-    #     print(f"SSH tunnel is open to {self.address}:{local_port}")
-
-    def restart_grpc_server(self, _rh_install_url=None, resync_rh=True, restart_ray=False):
-        # TODO how do we capture errors if this fails?
-        if resync_rh:
-            self.sync_runhouse_to_cluster(_install_url=_rh_install_url)
-        kill_proc_cmd = f'pkill -f "python3 -m runhouse.grpc_handler.unary_server"'
-        grpc_server_cmd = f'screen -dm python3 -m runhouse.grpc_handler.unary_server'
-        cmds = [kill_proc_cmd]
-        if restart_ray:
-            cmds.append('ray stop')
-            cmds.append('ray start --head')  # Need to set gpus or Ray will block on cpu-only clusters
-        cmds.append(grpc_server_cmd)
-
-        # If we need different commands for debian or ubuntu, we can use this:
-        # Need to get actual provider in case provider == 'cheapest'
-        # handle = sky.global_user_state.get_cluster_from_name(self.name)['handle']
-        # cloud_provider = str(handle.launched_resources.cloud)
-        # ubuntu_kill_proc_cmd = f'fuser -k {UnaryService.DEFAULT_PORT}/tcp'
-        # debian_kill_proc_cmd = "kill -9 $(netstat -anp | grep 50052 | grep -o '[0-9]*/' | sed 's+/$++')"
-        # f'kill -9 $(lsof -t -i:{UnaryService.DEFAULT_PORT})'
-        # kill_proc_at_port_cmd = debian_kill_proc_cmd if cloud_provider == 'GCP' \
-        #     else ubuntu_kill_proc_cmd
-
-        status_codes = self.run(commands=cmds,
-                                stream_logs=True,
-                                )
-        # As of 2022-27-Dec still seems we need this.
-        import time
-        time.sleep(2)
-        return status_codes
-
     @contextlib.contextmanager
     def pause_autostop(self):
         sky.autostop(self.name, idle_minutes=-1)
         yield
         sky.autostop(self.name, idle_minutes=self.autostop_mins)
 
-    def call_grpc(self, serialized_func):
-        # TODO check if this is actually avoiding creating a duplicate tunnel when creating one send after another
-        if not self.is_connected():
-            self.connect_grpc()
-
-        # TODO would be great to pause autostop here but it's too slow, about 2 seconds
-        return self.client.call_fn_remotely(message=serialized_func)
-
-    def is_connected(self):
-        return self.client is not None and self.client.is_connected()
-
-    def disconnect(self):
-        if self._grpc_tunnel:
-            self._grpc_tunnel.stop()
-        # if self.client:
-        #     self.client.shutdown()
-
     # ----------------- SSH Methods ----------------- #
-
-    @staticmethod
-    def cluster_ssh_key(path_to_file):
-        try:
-            f = open(path_to_file, 'r')
-            private_key = f.read()
-            return private_key
-        except FileNotFoundError:
-            raise Exception(f'File with ssh key not found in: {path_to_file}')
-
-    @staticmethod
-    def path_to_cluster_ssh_key(path_to_file) -> str:
-        user_path = Path(path_to_file).expanduser()
-        return str(user_path)
 
     def ssh_creds(self):
         if not Path(self._yaml_path).exists():
@@ -533,105 +298,5 @@ class Cluster(Resource):
 
         return backend_utils.ssh_credential_from_yaml(self._yaml_path)
 
-    def rsync(self, source, dest, up, contents=False):
-        """ Note that ending `source` with a slash will copy the contents of the directory into dest,
-        while omitting it will copy the directory itself (adding a directory layer)."""
-        # FYI, could be useful: https://github.com/gchamon/sysrsync
-        if contents:
-            source = source + '/'
-            dest = dest + '/'
-        ssh_credentials = self.ssh_creds()
-        runner = command_runner.SSHCommandRunner(self.address, **ssh_credentials)
-        runner.rsync(source, dest, up=up, stream_logs=False)
-
     def ssh(self):
         subprocess.run(["ssh", f"{self.name}"])
-
-    def run(self, commands: list, stream_logs=True, port_forward=None, require_outputs=True):
-        """ Run a list of shell commands on the cluster. """
-        # TODO add name parameter to create Run object, and use sky.exec (after updating to sky 2.0):
-        # sky.exec(commands, cluster_name=self.name, stream_logs=stream_logs, detach=False)
-        runner = command_runner.SSHCommandRunner(self.address, **self.ssh_creds())
-        return_codes = []
-        for command in commands:
-            logger.info(f"Running command on {self.name}: {command}")
-            ret_code = runner.run(command,
-                                  require_outputs=require_outputs,
-                                  stream_logs=stream_logs,
-                                  port_forward=port_forward)
-            return_codes.append(ret_code)
-        return return_codes
-
-    def run_python(self, commands: list, stream_logs=True, port_forward=None):
-        """ Run a list of python commands on the cluster. """
-        command_str = '; '.join(commands)
-        self.run([f'python3 -c "{command_str}"'], stream_logs=stream_logs, port_forward=port_forward)
-
-    def send_secrets(self, reload=False, providers: Optional[List[str]] = None):
-        if providers is not None:
-            # Send secrets for specific providers from local configs rather than trying to load from Vault
-            from runhouse import Secrets
-            secrets: list = Secrets.load_provider_secrets(providers=providers)
-            # TODO [JL] change this API so we don't have to convert the list to a dict
-            secrets: dict = {s['provider']: {k: v for k, v in s.items() if k != 'provider'} for s in secrets}
-            load_secrets_cmd = ['import runhouse as rh',
-                                f'rh.Secrets.save_provider_secrets(secrets={secrets})']
-        elif not self._secrets_sent or reload:
-            load_secrets_cmd = ['import runhouse as rh',
-                                'rh.Secrets.download_into_env()']
-        else:
-            # Secrets already sent and not reloading
-            return
-
-        self.run_python(load_secrets_cmd, stream_logs=True)
-        # TODO [JL] change this to a list to make sure new secrets get sent when the user wants to
-        self._secrets_sent = True
-
-    def ipython(self):
-        # TODO tunnel into python interpreter in cluster
-        pass
-
-    def notebook(self, persist=False, sync_package_on_close=None, port_forward=8888):
-        tunnel, port_fwd = self.ssh_tunnel(local_port=port_forward, num_ports_to_try=10)
-        try:
-            install_cmd = "pip install jupyterlab"
-            jupyter_cmd = f'jupyter lab --port {port_fwd} --no-browser'
-            # port_fwd = '-L localhost:8888:localhost:8888 '  # TOOD may need when we add docker support
-            with self.pause_autostop():
-                self.run(commands=[install_cmd, jupyter_cmd], stream_logs=True)
-
-        finally:
-            if sync_package_on_close:
-                if sync_package_on_close == './':
-                    sync_package_on_close = rns_client.locate_working_dir()
-                pkg = Package.from_string('local:' + sync_package_on_close)
-                self.rsync(source=f'~/{pkg.name}', dest=pkg.local_path, up=False)
-            if not persist:
-                tunnel.stop(force=True)
-                kill_jupyter_cmd = f'jupyter notebook stop {port_fwd}'
-                self.run(commands=[kill_jupyter_cmd])
-
-
-# Cluster factory method
-def cluster(name: str,
-            instance_type: str = None,
-            num_instances: int = None,
-            provider: str = None,
-            autostop_mins: int = None,
-            dryrun: bool = False,
-            use_spot: bool = None,
-            image_id: str = None,
-            ) -> Cluster:
-    config = rns_client.load_config(name)
-    config['name'] = name or config.get('rns_address', None) or config.get('name')
-
-    config['instance_type'] = instance_type or config.get('instance_type', None)
-    config['num_instances'] = num_instances or config.get('num_instances', None)
-    config['provider'] = provider or config.get('provider', None)
-    config['autostop_mins'] = autostop_mins if autostop_mins is not None else config.get('autostop_mins', None)
-    config['use_spot'] = use_spot if use_spot is not None else config.get('use_spot', None)
-    config['image_id'] = image_id if image_id is not None else config.get('image_id', None)
-
-    new_cluster = Cluster.from_config(config, dryrun=dryrun)
-
-    return new_cluster
