+++ runhouse/rns/tables/ray_table.py
@@ -1,7 +1,7 @@
 from typing import Optional, List
 
 from .table import Table
-from .. import Cluster
+from .. import SkyCluster
 from ..top_level_rns_fns import save
 
 
@@ -15,7 +15,7 @@ class RayTable(Table):
     def from_config(config: dict, dryrun=True):
         """ Load config values into the object. """
         if isinstance(config['fs'], dict):
-            config['fs'] = Cluster.from_config(config['fs'], dryrun=dryrun)
+            config['fs'] = SkyCluster.from_config(config['fs'], dryrun=dryrun)
         return RayTable(**config)
 
     def save(self,
