+++ runhouse/rns/top_level_rns_fns.py
@@ -136,13 +136,13 @@ def get_pinned_object(key: str, default=None):
 
 
 def get(key: str, cluster=None, default=None):
-    from runhouse.rns.hardware.skycluster import Cluster
+    from runhouse.rns.hardware.skycluster import SkyCluster
     if isinstance(cluster, str):
         if cluster == rh_config.obj_store.cluster_name:
             # We're currently on cluster, so just get the object from local rh_config.obj_store
             return rh_config.obj_store.get(key, default=default)
         else:
-            cluster = Cluster.from_name(cluster)
+            cluster = SkyCluster.from_name(cluster)
 
     if cluster.name == rh_config.obj_store.cluster_name:
         return rh_config.obj_store.get(key, default=default)
